refactor: 拆分音频引擎 NAPI 绑定
